fix(api): break the two real import cycles
north-star-architecture.md Phase 0: admin.ts <-> job-admin-routes.ts
and agent-channel.ts <-> stage-hints.ts <-> staged-preview.ts <->
agent-channel.ts were both real (value-level) cycles with no home for
the shared primitive they fought over.

- Move isAdmin/isAdminSession out of admin.ts into a new
  admin-session.ts; update the seven importers. job-admin-routes.ts no
  longer imports from admin.ts at all, so that cycle is gone.
- Move MAX_BUILD_PREVIEW_BYTES out of agent-channel.ts into a new
  build-preview-limits.ts; staged-preview.ts now imports the constant
  from there instead of from agent-channel.ts. Also fixes a duplicate
  hardcoded 320 * 1024 in submissions.ts that had drifted into its own
  local constant.
- Reseal the comment-prose baseline: the moved doc comments are
  rewritten as short one-liners in the new files (new files get zero
  prose debt), and the source files shrink accordingly.
