+++ apps/api/src/auth.ts
@@ -4,7 +4,7 @@ import type { FastifyInstance, FastifyReply, FastifyRequest } from 'fastify';
 import { OAuth2Client } from 'google-auth-library';
 import { z } from 'zod';
 import { resolveAccessTokenUser } from './access-token-service.js';
-import { isAdmin, isAdminSession } from './admin.js';
+import { isAdmin, isAdminSession } from './admin-session.js';
 import { isReviewer, isReviewerSession } from './review.js';
 import { resolveAppleAccount } from './apple-account.js';
 import { createAppleAuthVerifierFromEnv, parseAppleClientIds, type AppleAuthVerifier } from './apple-auth.js';
