+++ apps/api/src/build-preview-limits.ts
@@ -1,0 +1,2 @@
+// 320KB decoded — clears the games-repo bundle cap; never duplicate this number.
+export const MAX_BUILD_PREVIEW_BYTES = 320 * 1024;
