+++ eslint-rules/comment-prose-baseline.json
@@ -23,16 +23,16 @@
     "apps/api/src/access-token.ts": 509,
     "apps/api/src/account-deletion-routes.ts": 10,
     "apps/api/src/admin.test.ts": 323,
-    "apps/api/src/admin.ts": 2027,
-    "apps/api/src/agent-backend-env.ts": 252,
-    "apps/api/src/agent-backend.ts": 667,
+    "apps/api/src/admin.ts": 1938,
+    "apps/api/src/agent-backend-env.ts": 168,
+    "apps/api/src/agent-backend.ts": 654,
     "apps/api/src/agent-build-brief.test.ts": 21,
     "apps/api/src/agent-build-brief.ts": 95,
     "apps/api/src/agent-build-examples.ts": 103,
     "apps/api/src/agent-build-media.test.ts": 667,
     "apps/api/src/agent-build-reads.test.ts": 14,
-    "apps/api/src/agent-channel.test.ts": 1239,
-    "apps/api/src/agent-channel.ts": 4359,
+    "apps/api/src/agent-channel.test.ts": 1226,
+    "apps/api/src/agent-channel.ts": 4265,
     "apps/api/src/agent-creator-key-resolve.ts": 172,
     "apps/api/src/agent-creator-key.test.ts": 15,
     "apps/api/src/agent-creator-key.ts": 257,
@@ -117,7 +117,7 @@
     "apps/api/src/game-saves.test.ts": 91,
     "apps/api/src/game-saves.ts": 495,
     "apps/api/src/game-seed.test.ts": 416,
-    "apps/api/src/game-seed.ts": 1774,
+    "apps/api/src/game-seed.ts": 1735,
     "apps/api/src/game-snapshot-publish.test.ts": 201,
     "apps/api/src/game-snapshot-publish.ts": 431,
     "apps/api/src/game-snapshot-serve.test.ts": 272,
@@ -169,7 +169,7 @@
     "apps/api/src/mcp-oauth-metadata.test.ts": 73,
     "apps/api/src/mcp-oauth-metadata.ts": 319,
     "apps/api/src/mcp-open-round.test.ts": 71,
-    "apps/api/src/mcp-presence.ts": 224,
+    "apps/api/src/mcp-presence.ts": 212,
     "apps/api/src/mcp-server-discovery.test.ts": 90,
     "apps/api/src/mcp-server-discovery.ts": 316,
     "apps/api/src/mcp-server.test.ts": 1547,
@@ -208,7 +208,7 @@
     "apps/api/src/openai-apps-challenge.test.ts": 62,
     "apps/api/src/openai-apps-challenge.ts": 181,
     "apps/api/src/operator-alerts.test.ts": 232,
-    "apps/api/src/operator-alerts.ts": 1200,
+    "apps/api/src/operator-alerts.ts": 1199,
     "apps/api/src/owner-games.ts": 98,
     "apps/api/src/owner-of-record.ts": 429,
     "apps/api/src/player-feedback.test.ts": 72,
@@ -241,7 +241,7 @@
     "apps/api/src/refine.test.ts": 178,
     "apps/api/src/refine.ts": 761,
     "apps/api/src/remix-debug-flag.test.ts": 155,
-    "apps/api/src/remix-save.ts": 260,
+    "apps/api/src/remix-save.ts": 180,
     "apps/api/src/remix-suggestions.test.ts": 63,
     "apps/api/src/remix-suggestions.ts": 308,
     "apps/api/src/remix-turns.ts": 98,
@@ -254,7 +254,7 @@
     "apps/api/src/seed-bundle.ts": 302,
     "apps/api/src/seed-context.test.ts": 142,
     "apps/api/src/seed-context.ts": 574,
-    "apps/api/src/seed-status.ts": 58,
+    "apps/api/src/seed-status.ts": 13,
     "apps/api/src/self-build-backend.ts": 263,
     "apps/api/src/self-build-connect.test.ts": 45,
     "apps/api/src/self-build-connect.ts": 507,
@@ -274,11 +274,11 @@
     "apps/api/src/static-serving.test.ts": 70,
     "apps/api/src/store-firestore.test.ts": 566,
     "apps/api/src/store.test.ts": 486,
-    "apps/api/src/store.ts": 16333,
+    "apps/api/src/store.ts": 16323,
     "apps/api/src/submission-status.test.ts": 35,
     "apps/api/src/submission-status.ts": 1815,
     "apps/api/src/submissions.test.ts": 4563,
-    "apps/api/src/submissions.ts": 14559,
+    "apps/api/src/submissions.ts": 14543,
     "apps/api/src/suggestion-inbox.test.ts": 538,
     "apps/api/src/suggestion-inbox.ts": 791,
     "apps/api/src/suggestion-outcomes.test.ts": 154,
@@ -305,8 +305,8 @@
     "apps/api/src/verify-erase-signals.ts": 1140,
     "apps/api/src/visit-funnel.test.ts": 308,
     "apps/api/src/visit-funnel.ts": 1962,
-    "apps/api/src/visit-telemetry.test.ts": 156,
-    "apps/api/src/visit-telemetry.ts": 743,
+    "apps/api/src/visit-telemetry.test.ts": 121,
+    "apps/api/src/visit-telemetry.ts": 732,
     "apps/api/src/votes.test.ts": 50,
     "apps/api/src/votes.ts": 175,
     "apps/api/src/workspace-archive.test.ts": 98,
@@ -371,7 +371,7 @@
     "apps/web/src/ClaimHandleModal.tsx": 67,
     "apps/web/src/ClosedBetaSplash.tsx": 146,
     "apps/web/src/CodeSurface.test.tsx": 114,
-    "apps/web/src/CodeSurface.tsx": 647,
+    "apps/web/src/CodeSurface.tsx": 630,
     "apps/web/src/codeSurfaceApi.ts": 82,
     "apps/web/src/codeTokens.ts": 177,
     "apps/web/src/connectApi.ts": 231,
@@ -394,13 +394,13 @@
     "apps/web/src/CreatorStudioView.handoff.test.tsx": 255,
     "apps/web/src/CreatorStudioView.test.tsx": 858,
     "apps/web/src/CreatorStudioView.theater.test.tsx": 69,
-    "apps/web/src/CreatorStudioView.tsx": 2976,
+    "apps/web/src/CreatorStudioView.tsx": 2968,
     "apps/web/src/editorBridge.ts": 228,
     "apps/web/src/EditorCollections.test.tsx": 13,
     "apps/web/src/editorContentTools.ts": 147,
     "apps/web/src/editorJsonLiveDiff.ts": 92,
     "apps/web/src/EditorPanel.test.tsx": 16,
-    "apps/web/src/EditorPanel.tsx": 645,
+    "apps/web/src/EditorPanel.tsx": 644,
     "apps/web/src/EditProfileModal.tsx": 56,
     "apps/web/src/GameDetailPage.test.tsx": 15,
     "apps/web/src/GameDetailPage.tsx": 36,
@@ -412,7 +412,7 @@
     "apps/web/src/gamePageApi.ts": 59,
     "apps/web/src/gamePlayer.bridge.test.ts": 334,
     "apps/web/src/gamePlayer.hiddenFrame.test.tsx": 97,
-    "apps/web/src/gamePlayer.ts": 1603,
+    "apps/web/src/gamePlayer.ts": 1597,
     "apps/web/src/gamePlayerBridge.test.ts": 528,
     "apps/web/src/gameSave.test.tsx": 281,
     "apps/web/src/gameSave.ts": 756,
@@ -431,7 +431,7 @@
     "apps/web/src/handVerbs.ts": 309,
     "apps/web/src/headerMascot.test.ts": 279,
     "apps/web/src/healthApi.ts": 453,
-    "apps/web/src/HeroPromptSection.tsx": 42,
+    "apps/web/src/HeroPromptSection.tsx": 35,
     "apps/web/src/howToPlay.test.ts": 329,
     "apps/web/src/howToPlay.ts": 902,
     "apps/web/src/HowToPlayPanel.test.tsx": 175,
@@ -465,8 +465,8 @@
     "apps/web/src/mp/roomClient.ts": 92,
     "apps/web/src/mp/voicePhrases.ts": 13,
     "apps/web/src/mySpecs.ts": 26,
-    "apps/web/src/NavHeader.test.tsx": 178,
-    "apps/web/src/NavHeader.tsx": 366,
+    "apps/web/src/NavHeader.test.tsx": 165,
+    "apps/web/src/NavHeader.tsx": 351,
     "apps/web/src/NotFoundPage.tsx": 33,
     "apps/web/src/NotificationBell.test.tsx": 119,
     "apps/web/src/NotificationBell.tsx": 187,
@@ -523,14 +523,13 @@
     "apps/web/src/SiteFooter.test.tsx": 146,
     "apps/web/src/SiteFooter.tsx": 214,
     "apps/web/src/splashMascotPullups.test.ts": 34,
-    "apps/web/src/SplitHero.tsx": 10,
-    "apps/web/src/studioApi.ts": 835,
+    "apps/web/src/studioApi.ts": 818,
     "apps/web/src/StudioChatRail.tsx": 230,
     "apps/web/src/StudioConnectCard.test.tsx": 66,
     "apps/web/src/StudioConnectCard.tsx": 472,
     "apps/web/src/StudioCreatorAgentKeyPanel.tsx": 37,
     "apps/web/src/studioCreatorProfile.tsx": 13,
-    "apps/web/src/StudioDetailsBuildProgress.tsx": 58,
+    "apps/web/src/StudioDetailsBuildProgress.tsx": 14,
     "apps/web/src/StudioDetailsMedia.tsx": 50,
     "apps/web/src/StudioFullBleed.tsx": 46,
     "apps/web/src/StudioOAuthClientsPanel.tsx": 21,
@@ -554,7 +553,7 @@
     "apps/web/src/styles.studioShell.test.ts": 42,
     "apps/web/src/submissionApi.ts": 1241,
     "apps/web/src/SubmissionStatusView.test.ts": 1780,
-    "apps/web/src/SubmissionStatusView.tsx": 4118,
+    "apps/web/src/SubmissionStatusView.tsx": 4092,
     "apps/web/src/SuggestionsPanel.test.tsx": 28,
     "apps/web/src/SuggestionsPanel.tsx": 113,
     "apps/web/src/telemetry.test.ts": 201,
@@ -582,7 +581,7 @@
     "apps/web/src/VisitFunnelPanel.tsx": 407,
     "apps/web/src/visitTelemetry.dom.test.ts": 185,
     "apps/web/src/visitTelemetry.test.ts": 170,
-    "apps/web/src/visitTelemetry.ts": 2645,
+    "apps/web/src/visitTelemetry.ts": 2640,
     "apps/web/src/viteEnvWiring.test.ts": 435,
     "apps/web/src/voiceMeter.ts": 153,
     "apps/web/src/votesApi.ts": 57,
